pipelines/requirements.txt
requests>=2.31.0
polars>=0.20.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
click>=8.1.0
